Fixed reports from showing '_' between each character
The system was printing reports with crazy long names to to how it
was splitting the list of domains and device prefixes for printing
reports. To fix this I added a new variable for each that handles the
formatting needs to get the report to show the correct way and used
the logging module to look over what was occurring.